socialapi: fix code quality issues
