+++ go/src/socialapi/workers/algoliaconnector/algoliaconnector/account_test.go
@@ -15,6 +15,26 @@ import (
 	. "github.com/smartystreets/goconvey/convey"
 )
 
+// makeSureAccount checks if the given id's get request returns the desired err,
+// it will re-try every 100ms until deadline of 2 minutes reached. Algolia
+// doesnt index the records right away, so try to go to a desired state
+func makeSureAccount(handler *Controller, id string, f func(map[string]interface{}, error) bool) error {
+	deadLine := time.After(time.Minute * 2)
+	tick := time.Tick(time.Millisecond * 100)
+	for {
+		select {
+		case <-tick:
+			record, err := handler.get(IndexAccounts, id)
+			if f(record, err) {
+				return nil
+			}
+		case <-deadLine:
+			handler.log.Critical("deadline reached on account but not returning an error")
+			return nil
+		}
+	}
+}
+
 func TestAccountTesting(t *testing.T) {
 	runner, handler := getTestHandler()
 	defer runner.Close()
