+++ go/src/socialapi/workers/algoliaconnector/algoliaconnector/helpers.go
@@ -79,54 +79,3 @@ func (f *Controller) tagAction(action, indexName string, objectID, value interfa
 		},
 	})
 }
-
-func appendTag(record map[string]interface{}, channelId string) []interface{} {
-	if record == nil {
-		return []interface{}{channelId}
-	}
-
-	tagsDoc, ok := record["_tags"]
-	if !ok {
-		return []interface{}{channelId}
-	}
-
-	tags, tagsOk := tagsDoc.([]interface{})
-	if !tagsOk {
-		return []interface{}{channelId}
-	}
-	for _, ele := range tags {
-		if ele == channelId {
-			return tags
-		}
-	}
-
-	return append(tags, channelId)
-}
-
-func removeMessageTag(record map[string]interface{}, channelId string) []interface{} {
-	if record == nil {
-		return nil
-	}
-
-	filtered := make([]interface{}, 0)
-
-	tagsDoc, ok := record["_tags"]
-	if !ok {
-		return nil
-	}
-
-	tags, tagsOk := tagsDoc.([]interface{})
-	if !tagsOk {
-		return nil
-	}
-
-	for _, ele := range tags {
-		id := ele.(string)
-		if id == channelId {
-			continue
-		}
-		filtered = append(filtered, id)
-	}
-
-	return filtered
-}
