+++ go/src/socialapi/workers/api/modules/reply/handlers.go
@@ -1,7 +1,6 @@
 package reply
 
 import (
-	"socialapi/models"
 	"socialapi/workers/common/handler"
 	"socialapi/workers/common/mux"
 )
@@ -14,7 +13,6 @@ func AddHandlers(m *mux.Mux) {
 			Name:     "reply-create",
 			Type:     handler.PostRequest,
 			Endpoint: "/message/{id}/reply",
-			Securer:  models.MessageSecurer,
 		},
 	)
 
@@ -24,7 +22,6 @@ func AddHandlers(m *mux.Mux) {
 			Name:     "reply-list",
 			Type:     handler.GetRequest,
 			Endpoint: "/message/{id}/reply",
-			Securer:  models.MessageReadSecurer,
 		},
 	)
 }
